backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt==3.2.0",
    "fastapi>=0.116.1",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.5.0",
    "python-mpd2>=3.1.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.42",
    "uvicorn>=0.35.0",
]
